requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.16.0
colorama==0.4.6
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
openai==1.40.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygit==0.1
pygit2==1.15.1
PyYAML==6.0.2
smmap==5.0.1
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
